% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.col.R
\name{add.col}
\alias{add.col}
\title{add column}
\usage{
add.col(dataf, vec, namevec)
}
\arguments{
\item{dataf}{dataframe}

\item{vec}{vector}

\item{namevec}{the names of the columns of vector}
}
\value{
dataf dataframe combined with the vector
}
\description{
take a data-frame and a vector and combine by columns, respectively.
}
