% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atualizaCampoBase.R
\name{atualizaCampoBase}
\alias{atualizaCampoBase}
\title{updated base field}
\usage{
atualizaCampoBase(camposAtualizar, baseAgrupada, baseAtualizar, keys,
  verbose = FALSE)
}
\arguments{
\item{camposAtualizar}{is the vector you want to update}

\item{baseAgrupada}{It is the database that contains the data you want to update on dataframe}

\item{baseAtualizar}{It is dataframe that you want to change fields}

\item{keys}{are the keys of the table that will be used in the compare}

\item{verbose}{default false}
}
\value{
baseAtualizar with the updated fields according to baseAgrupada
}
\description{
this function update certain fields in a dataframe, based on the provided key
}

