% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avaliaAjuste.R
\name{avaliaAjuste}
\alias{avaliaAjuste}
\title{avalia Ajuste}
\usage{
avaliaAjuste(dataFrame, variavelObservados, variavelEstimados,
  linear = TRUE, nParametros = NA, intercepto = TRUE, plot = NA,
  modelo = NA, resumo = FALSE, emf = TRUE)
}
\arguments{
\item{dataFrame}{dataFrane with information observed, estimated}

\item{variavelObservados}{vector of values observed.}

\item{variavelEstimados}{vector of values estimated.}

\item{linear}{boolean is linear model}

\item{nParametros}{number of parameters used in the adjusted model}

\item{intercepto}{if you model is no-intercepto use FALSE}

\item{plot}{Vector graphic information}

\item{modelo}{the name of the adjusted model}

\item{resumo}{if you want summary information, use TRUE}

\item{emf}{to save the graphic in the format emf use TRUE}
}
\description{
this function evaluates the quality of the adjustment of the statistical model, rom observed data and those estimated by the model, observed
}
