% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avaliaEstimativas.R
\name{avaliaEstimativas}
\alias{avaliaEstimativas}
\title{calculate Estimates}
\usage{
avaliaEstimativas(observado, estimado, estatisticas, ajuste = NULL,
  graficos = NULL, salvarEm = NULL, nome = "observadoXestimado")
}
\arguments{
\item{observado}{list containing the observations of variable}

\item{estimado}{list containing estimates of variable}

\item{estatisticas}{list of arg to calc estatistics}

\item{ajuste}{is ajust obtained a function like lm or nlsLM}

\item{graficos}{list of arg to plot graphics}

\item{salvarEm}{directory to save files}

\item{nome}{name of files will be save}
}
\value{
will be returned
}
\description{
given a list of observations and an estimated list of these observations this function evaluates how close it is the estimated value of observed and saves the differences
}

