% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avaliaVolumeAgeBased.R
\name{avaliaVolumeAgeBased}
\alias{avaliaVolumeAgeBased}
\title{avalia Volume Age Based}
\usage{
avaliaVolumeAgeBased(base, firstAge, lastAge, models, mapper = list(age1
  = "idade1", age2 = "idade2", dap1 = "dap1", dap2 = "dap2", dap2est =
  "dap2est", ht1 = "ht1", ht2 = "ht2", ht2est = "ht2est", volume1 =
  "volume1", volume2 = "volume2", volume2est = "volume2est"),
  groupBy = "parcela", save = NULL, percTraining = 0.7,
  paramEstatisticsDAP, paramEstatisticsHT, paramEstatisticsVolume,
  plot = "parcela", ageER = "^.*_", ageRound = NaN, ageInYears = F,
  forcePredict = F)
}
\arguments{
\item{base}{the data base}

\item{firstAge}{the first age to eval}

\item{lastAge}{the last age to eval}

\item{models}{list of exclusive for base models}

\item{mapper}{mapper from labels of fields volume, dap, ht}

\item{groupBy}{name field of base is group of individuals}

\item{save}{list of param to save the files}

\item{percTraining}{percentage that will be reserved for training (default 0.70)}

\item{paramEstatisticsDAP}{parameters to pass to function 'fnAvaliaEstimativas'}

\item{paramEstatisticsHT}{analogous to paramEstatisticsDAP}

\item{paramEstatisticsVolume}{analogous to paramEstatisticsDAP}

\item{plot}{is list of plots to function roundAges}

\item{ageER}{regex used to discover age in names from dataframe in listOfdata}

\item{ageRound}{synchronize begin of ages with an age? what age?}

\item{ageInYears}{ages are in year?}

\item{forcePredict}{force the calculation without using predict?}
}
\value{
will be returned a list of round ages
}
\description{
this function evaluate volume based on ages
}
