% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avaliaVolumeAvancado.R
\name{avaliaVolumeAvancado}
\alias{avaliaVolumeAvancado}
\title{evaluates Volume Advanced}
\usage{
avaliaVolumeAvancado(base, mapeamento = list(dap1 = "dap1", dap2 = "dap2", ht1
  = "ht1", ht2 = "ht2"), modelos = NULL, salvar = NULL, graficos = NULL,
  estatisticas = NULL, forcePredict = F, dividirEm = "parcela",
  percentualDeTreino = 0.7, agruparPor = "parcela",
  fnCalculaVolume = calculaVolumeDefault)
}
\arguments{
\item{base}{data.frame with data}

\item{mapeamento}{name of field eight and diameter}

\item{modelos}{list of exclusive for base models}

\item{salvar}{list of param to save the files}

\item{graficos}{list of param to plot graphics}

\item{estatisticas}{list of param to caclc estatistics}

\item{forcePredict}{force the calculation without using predict?}

\item{dividirEm}{how divide the base in training and validation}

\item{percentualDeTreino}{how many percent will stay in the training group?}

\item{agruparPor}{name field of base is group of individuals}

\item{fnCalculaVolume}{list of estatistics results}
}
\value{
will be returned a result of statistics and ranking of volume
}
\description{
this function performs an assessment of estimates of a variable as the forcefulness with expected
}

