% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculaA.R
\name{calculaA}
\alias{calculaA}
\title{Fator A}
\usage{
calculaA(n, k)
}
\arguments{
\item{n}{the size of the vector of regression model data}

\item{k}{is the number of model parameters}
}
\description{
The linear intercept model,
}
\details{
a = (n-1)/(n-k-1)
}

