% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculaPerc.R
\name{calculaPerc}
\alias{calculaPerc}
\title{calculates percentage}
\usage{
calculaPerc(valor, observados)
}
\arguments{
\item{valor}{number amount you to know the percentage}

\item{observados}{number relationship to which you want to calculate the percentage, if it is a vector of integers is calculated its average.}
}
\description{
With this function, you can calculate the ratio of one quantity or magnitude relative to another evaluated in percentage.
}
\details{
calculaPerc = ((valor)/mean(observados))*100
}

