% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculaVolumeDefault.R
\name{calculaVolumeDefault}
\alias{calculaVolumeDefault}
\title{calculates Volume Default}
\usage{
calculaVolumeDefault(ht, dap, ...)
}
\arguments{
\item{ht}{is list of height of individuals}

\item{dap}{is list of diameter of individuals}

\item{...}{only for compatibility with other functions}
}
\value{
will be returned a list of volume calc
}
\description{
this function calculates the volume based on the height and volume of literature of the equation
}

