% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce.R
\name{ce}
\alias{ce}
\title{coefficient of efficiency}
\usage{
ce(observados, estimados)
}
\arguments{
\item{observados}{vector of values observed.}

\item{estimados}{vector of regression model data.}
}
\description{
Nash Sutcliffe 1970 model efficiency coefficient is used to assess the predictive power of hydrological models.
}
\references{
( Nash and Sutcliffe, 1970) \url{https://en.wikipedia.org/wiki/Nash-Sutcliffe_model_efficiency_coefficient} for more details.
}

