% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificarDAP.R
\name{classificarDAP}
\alias{classificarDAP}
\title{classify field dap}
\usage{
classificarDAP(inventario, amplitude = 1, verbose = FALSE)
}
\arguments{
\item{inventario}{the database to update}

\item{amplitude}{it is amplitude of dap class}

\item{verbose}{use TRUE to show status of process}
}
\value{
data.frame with classeDAP field and other
}
\description{
classify field dap as specified amplitude and includes a few fields
}

