% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/criaDadosPareados.R
\name{criaDadosPareados}
\alias{criaDadosPareados}
\title{Create Date Paired}
\usage{
criaDadosPareados(dataFrame, campoChave, campoComparacao, camposPareados,
  camposNaoPareados)
}
\arguments{
\item{dataFrame}{dataframe that you want to pair dataFrame must contain columns cod_id, ANO_MEDICAO1, ANO_MEDICAO2, DAP1, DAP2, HT1, HT2, ID_PROJETO}

\item{campoChave}{character the column that will be paired}

\item{campoComparacao}{character the field used to compare the period of change}

\item{camposPareados}{vector the fields that will be paired exemple CampoesPareados=c(dap,ht)}

\item{camposNaoPareados}{the fields he wants to be present without the paired}
}
\value{
will be returned a dataframe containing columns
cod_id, ANO_MEDICAO1, ANO_MEDICAO2, DAP1, DAP2, HT1, HT2, ID_PROJETO
}
\description{
paired a dataframe
}

