% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineClasses2.R
\name{defineClasses2}
\alias{defineClasses2}
\title{define Classes 2}
\usage{
defineClasses2(dados, amplitude)
}
\arguments{
\item{dados}{a vector of numbers}

\item{amplitude}{integer Class amplitude range}
}
\description{
creates a list with the class interval of a frequency distribution
}
\examples{
dados <- c(1,2,3,4)
defineClasses2(dados,2)
}
