% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estatisticas.R
\name{estatisticas}
\alias{estatisticas}
\title{Estatistics}
\usage{
estatisticas(observado, estimado, dfEstatisticas = NULL, ...)
}
\arguments{
\item{observado}{list containing the observations of variable}

\item{estimado}{list containing estimates of variable}

\item{dfEstatisticas}{a data.frame}

\item{...}{only for compatibility with other functions}
}
\value{
will be returned a list with data.frame with observado and estimado fields and other with statictcs of model add
}
\description{
this function returns a data.frame containing fields observado and estimado
}
