% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estatisticasBIAS.R
\name{estatisticasBIAS}
\alias{estatisticasBIAS}
\title{BIAS Estatistics}
\usage{
estatisticasBIAS(observado, estimado, dfEstatisticas = NULL, ...)
}
\arguments{
\item{observado}{list containing the observations of variable}

\item{estimado}{list containing estimates of variable}

\item{dfEstatisticas}{a data.frame}

\item{...}{only for compatibility with other functions}
}
\value{
will be returned data.frame with bias
}
\description{
this function returns a data.frame containing fields bias
}

