% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estatisticasBiasPERCENTUAL.R
\name{estatisticasBiasPERCENTUAL}
\alias{estatisticasBiasPERCENTUAL}
\title{percent BIAS Estatistics}
\usage{
estatisticasBiasPERCENTUAL(observado, estimado, dfEstatisticas, ...)
}
\arguments{
\item{observado}{list containing the observations of variable}

\item{estimado}{list containing estimates of variable}

\item{dfEstatisticas}{a data.frame with field bias}

\item{...}{only for compatibility with other functions}
}
\value{
will be returned data.frame with biasPERCENTUAL
}
\description{
this function returns a data.frame containing fields biasPERCENTUAL
}
