% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estatisticasCorrPERCENTUAL.R
\name{estatisticasCorrPERCENTUAL}
\alias{estatisticasCorrPERCENTUAL}
\title{Percent Correlacion Estatistics}
\usage{
estatisticasCorrPERCENTUAL(observado, estimado, dfEstatisticas, ...)
}
\arguments{
\item{observado}{list containing the observations of variable}

\item{estimado}{list containing estimates of variable}

\item{dfEstatisticas}{a data.frame with corr field}

\item{...}{only for compatibility with other functions}
}
\value{
will be returned data.frame with corr_PERCENTUAL field
}
\description{
this function returns a data.frame containing fields corr_PERCENTUAL
}

