% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estatisticasCV.R
\name{estatisticasCV}
\alias{estatisticasCV}
\title{Co variance Estatistics}
\usage{
estatisticasCV(observado, estimado, ajuste = NULL,
  dfEstatisticas = NULL, baseDoAjuste = NULL, formulaDoAjuste = NULL,
  ...)
}
\arguments{
\item{observado}{list containing the observations of variable}

\item{estimado}{list containing estimates of variable}

\item{ajuste}{is ajust obtained a function like lm or nlsLM}

\item{dfEstatisticas}{a data.frame}

\item{baseDoAjuste}{data.frame optional}

\item{formulaDoAjuste}{formula used in ajust}

\item{...}{only for compatibility with other functions}
}
\value{
will be returned data.frame with cv
}
\description{
this function returns a data.frame containing fields cv
}
