% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estatisticasCvPERCENTUAL.R
\name{estatisticasCvPERCENTUAL}
\alias{estatisticasCvPERCENTUAL}
\title{Percent Co variance Estatistics}
\usage{
estatisticasCvPERCENTUAL(observado, estimado, dfEstatisticas, ...)
}
\arguments{
\item{observado}{list containing the observations of variable}

\item{estimado}{list containing estimates of variable}

\item{dfEstatisticas}{a data.frame with cv field}

\item{...}{only for compatibility with other functions}
}
\value{
will be returned data.frame with cvPERCENTUAL
}
\description{
this function returns a data.frame containing fields cvPERCENTUAL
}

