% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estatisticasR2.R
\name{estatisticasR2}
\alias{estatisticasR2}
\title{R2 Estatistics for linear models}
\usage{
estatisticasR2(observado, estimado, dfEstatisticas = NULL,
  ajuste = NULL, intercepto = TRUE, formulaDoAjuste = NULL,
  baseDoAjuste = NULL, ...)
}
\arguments{
\item{observado}{list containing the observations of variable}

\item{estimado}{list containing estimates of variable}

\item{dfEstatisticas}{a data.frame}

\item{ajuste}{is ajust obtained a function like lm or nlsLM}

\item{intercepto}{intercepts?}

\item{formulaDoAjuste}{formula used in ajust}

\item{baseDoAjuste}{data.frame optional}

\item{...}{only for compatibility with other functions}
}
\value{
will be returned data.frame with r2
}
\description{
this function returns a data.frame containing fields r2
}
