% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estatisticasResiduoPERCENTUAL.R
\name{estatisticasResiduoPERCENTUAL}
\alias{estatisticasResiduoPERCENTUAL}
\title{Residuals Estatistics}
\usage{
estatisticasResiduoPERCENTUAL(observado, estimado, dfEstatisticas = NULL, ...)
}
\arguments{
\item{observado}{list containing the observations of variable}

\item{estimado}{list containing estimates of variable}

\item{dfEstatisticas}{a data.frame containing field residuo}

\item{...}{only for compatibility with other functions}
}
\value{
will be returned data.frame with percent Residuals field
}
\description{
this function returns a data.frame containing field residuoPERCENTUAL
}

