% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estatisticasRMSE.R
\name{estatisticasRMSE}
\alias{estatisticasRMSE}
\title{RMSE Estatistics}
\usage{
estatisticasRMSE(observado, estimado, dfEstatisticas = NULL, ...)
}
\arguments{
\item{observado}{list containing the observations of variable}

\item{estimado}{list containing estimates of variable}

\item{dfEstatisticas}{a data.frame}

\item{...}{only for compatibility with other functions}
}
\value{
will be returned data.frame with RMSE calc
}
\description{
this function returns a data.frame containing fields rmse
}

