% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estatisticasRmsePERCENTUAL.R
\name{estatisticasRmsePERCENTUAL}
\alias{estatisticasRmsePERCENTUAL}
\title{percent RMSE Estatistics}
\usage{
estatisticasRmsePERCENTUAL(observado, estimado, dfEstatisticas, ...)
}
\arguments{
\item{observado}{list containing the observations of variable}

\item{estimado}{list containing estimates of variable}

\item{dfEstatisticas}{a data.frame containing field rmse}

\item{...}{only for compatibility with other functions}
}
\value{
will be returned data.frame with rmse PERCENTUAL calc
}
\description{
this function returns a data.frame containing fields rmsePERCENTUAL
}
