% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalAgeBased.R
\name{evalAgeBased}
\alias{evalAgeBased}
\title{Evaluate Age Based}
\usage{
evalAgeBased(listOfdata, mapper = list(volume2 = "volume2", volume2est =
  "volume2est", dap2 = "dap2", dap2est = "dap2est", ht2 = "ht2", ht2est =
  "ht2est"), fnAvaliaEstimativas = avaliaEstimativas, paramEstatisticsDAP,
  paramEstatisticsHT, paramEstatisticsVolume,
  titulos = "paste(\\"Idade\\", idade)", ageER = "^.*_", nameModel = NULL)
}
\arguments{
\item{listOfdata}{the list that contains the data frames predicts}

\item{mapper}{mapper from labels of fields volume, dap, ht}

\item{fnAvaliaEstimativas}{funcion to evaluate dataframes of listOfdata}

\item{paramEstatisticsDAP}{parameters to pass to function 'fnAvaliaEstimativas'}

\item{paramEstatisticsHT}{analogous to paramEstatisticsDAP}

\item{paramEstatisticsVolume}{analogous to paramEstatisticsDAP}

\item{titulos}{customize titles of grafics}

\item{ageER}{regex used to discover age in names from dataframe in listOfdata}

\item{nameModel}{name of model used to predict to generate listOfdata optional}
}
\value{
will be returned a list of round ages
}
\description{
This function evaluates the volume of past data frames based on the parameter 'listOfdata'
}

