% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fator_bias.R
\name{fator_bias}
\alias{fator_bias}
\title{Fator Bias}
\usage{
fator_bias(observados, estimados, n)
}
\arguments{
\item{observados}{vector of values observed.}

\item{estimados}{vector of values estimated.}

\item{n}{the size of the vector of regression model data}
}
\description{
The bias factor indicates the average of the observed values is above or below the equity line.
}
\details{
fator_bias = 10^(sum(log(estimados/observados)/n))
#'
}
\references{
see \url{http://smas.chemeng.ntua.gr/miram/files/publ_268_11_2_2005.pdf} for more details.
}

