% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geraModelo.R
\name{geraModelo}
\alias{geraModelo}
\title{Generates function to work with a model}
\usage{
geraModelo(nome = "modelo sem nome", formula,
  funcaoRegressao = "nlsLM", palpite = NULL, maisParametros = NULL,
  requires = NULL, customizado = NULL)
}
\arguments{
\item{nome}{is the name of model}

\item{formula}{is the string formula begin with y2~y1}

\item{funcaoRegressao}{is the function that will make the regression, ex.: 'nlsLM'}

\item{palpite}{param start of funcaoRegressao}

\item{maisParametros}{string add in funcaoRegressao, ex lm(y2~y1, data=base, maisParametros)}

\item{requires}{list of string of packges used to work with funcaoRegressao}

\item{customizado}{if you want to write as the return will be obtained report as a string}
}
\value{
will be returned a function with exclusive model
}
\description{
this function generates unique model given: A formula and a guess (optional: name, funcaoRegressao, maisParametros, requires - proidido: custom)] or[A string saying how the return will be obtained eg custom = "lm (dap2 dap1 ~ * b 0)" (if the formula can not be passed just go empty, ex .: formula = "")]
}
