% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getAnoMedicao.R
\name{getAnoMedicao}
\alias{getAnoMedicao}
\title{Get Year Measurement}
\usage{
getAnoMedicao(dataFrame)
}
\arguments{
\item{dataFrame}{that has the column DATA_MEDICAO,CD_PARCELA}
}
\value{
dataFrame dataframe that has columns COD_PARCELA, DATA_MEDICAO, ANO_MEDICAO
}
\description{
using DATA_MEDICAO column in the form MM/YYYY creates a new column with the name "ANO_MEDICAO" in YYYY format
}
\examples{
DATA_MEDICAO <- c("02/2009","02/2010","02/11","02/2011")
CD_PARCELA <- c(6947,6947,6947,6947)
test <- data.frame(DATA_MEDICAO,CD_PARCELA)
getAnoMedicao(test)
}

