% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBaseOfAjust.R
\name{getBaseOfAjust}
\alias{getBaseOfAjust}
\title{get database Of Ajust}
\usage{
getBaseOfAjust(ajuste)
}
\arguments{
\item{ajuste}{is ajust obtained a function like lm or nlsLM}
}
\value{
will be returned a string which is the database of ajust
}
\description{
this function returns the database used in the setting
}

