% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClasses.R
\name{getClasses}
\alias{getClasses}
\title{Get List of DAP Classes}
\usage{
getClasses(base, amplitude, verbose = FALSE)
}
\arguments{
\item{base}{the data.frame  containing fields limiteMin, limiteMax of parcela and idadearred}

\item{amplitude}{it is amplitude of dap class}

\item{verbose}{use TRUE to show status of process}
}
\value{
list of data.frame
}
\description{
this function return a list of data.frame where each contains a number of dap classes according to reported basis
}

