% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColumnsOfAjust.R
\name{getColumnsOfAjust}
\alias{getColumnsOfAjust}
\title{get Columns used in Ajust}
\usage{
getColumnsOfAjust(ajuste, dfDados = NULL, excludeY1andY2 = T)
}
\arguments{
\item{ajuste}{is ajust obtained a function like lm or nlsLM}

\item{dfDados}{data.frame optional}

\item{excludeY1andY2}{delete Y1 and Y2 fields? del formula(y1~y2...)}
}
\value{
will be returned list of columns used in ajust
}
\description{
this function returns an array with the column names that are on the model and reported basis or basis used in ajust
}

