% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColumnsOfBase.R
\name{getColumnsOfBase}
\alias{getColumnsOfBase}
\title{get Columns Of Base present in the string}
\usage{
getColumnsOfBase(base, strColumns)
}
\arguments{
\item{base}{data.frame}

\item{strColumns}{string containing name fields of the base}
}
\value{
will be returned list with fields whose name are present in the string
}
\description{
this function returns the columns of a base whose names are present in the string strColumns
}

