% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFormulaExclusivaOfAjust.R
\name{getFormulaExclusivaOfAjust}
\alias{getFormulaExclusivaOfAjust}
\title{get Formula Exclusive Of Ajust}
\usage{
getFormulaExclusivaOfAjust(ajuste)
}
\arguments{
\item{ajuste}{is ajust obtained a function like lm or nlsLM}
}
\value{
will be returned a string which is the formula of ajust
}
\description{
this function returns the formula of the model used in ajust
}

