% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGraphicHistogram.R
\name{getGraphicHistogram}
\alias{getGraphicHistogram}
\title{Get Histogram of Residuals absolute}
\usage{
getGraphicHistogram(titulo = "residuos", nome = "observadoXestimado",
  estatisticas, save = NULL, vetorial = T, ...)
}
\arguments{
\item{titulo}{is the title graphic}

\item{nome}{name of file case save}

\item{estatisticas}{data.frame containing field 'residuo'}

\item{save}{If you want to save enter the directory as a string}

\item{vetorial}{save picture in vector type? (Default TRUE)}

\item{...}{only for compatibility with other functions}
}
\description{
this function displays/saves a histogram graph illustrating the frequency of waste in classes
}

