% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGraphicObservadoXEstimado.R
\name{getGraphicObservadoXEstimado}
\alias{getGraphicObservadoXEstimado}
\title{Get Graphic Observed X Estimated}
\usage{
getGraphicObservadoXEstimado(titulo = "observadoXestimado",
  nome = "observadoXestimado", observado, estimado, showTestF = TRUE,
  save = NULL, labsX = "observado", labsy = "estimado", vetorial = T,
  ...)
}
\arguments{
\item{titulo}{is the title graphic}

\item{nome}{name of file case save}

\item{observado}{list containing the observations of variable}

\item{estimado}{list containing estimates of variable}

\item{showTestF}{draw results of test F in graphic?}

\item{save}{If you want to save enter the directory as a string}

\item{labsX}{label x}

\item{labsy}{label y}

\item{vetorial}{save picture in vector type? (Default TRUE)}

\item{...}{only for compatibility with other functions}
}
\description{
this function display/save a graphic scatter.smooth illustrating the difference between the observed and estimated
}

