% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGraphicResiduoPerc.R
\name{getGraphicResiduoPerc}
\alias{getGraphicResiduoPerc}
\title{Get Graphic Residuals percent}
\usage{
getGraphicResiduoPerc(titulo = "Residuo Percentual (\%)",
  nome = "observadoXestimado", strVariavelXResiduo = NULL,
  estatisticas, save = NULL, labsX = "observacao",
  labsy = "residuos", vetorial = T, ...)
}
\arguments{
\item{titulo}{is the title graphic}

\item{nome}{name of file case save}

\item{strVariavelXResiduo}{list containing variable for compare with residuals}

\item{estatisticas}{data.frame containing field 'residuoPERCENTUAL'}

\item{save}{If you want to save enter the directory as a string}

\item{labsX}{label x}

\item{labsy}{label y}

\item{vetorial}{save picture in vector type? (Default TRUE)}

\item{...}{only for compatibility with other functions}
}
\description{
this function displays/saves a graph illustrating the distribution scatter.smooth of residues
}
