% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParametrosOfModel.R
\name{getParametrosOfModel}
\alias{getParametrosOfModel}
\title{get Parametros Of Model}
\usage{
getParametrosOfModel(ajuste, base = NULL, formula = NULL)
}
\arguments{
\item{ajuste}{is ajust obtained a function like lm or nlsLM}

\item{base}{optional data.frame whose fields name is present in formula}

\item{formula}{string containing name fields of the base}
}
\value{
will be returned list of columns used in ajust or in formula
}
\description{
this function retona columns the base of the parameter or setting present in the model
}
