% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravaDocResultado.R
\name{gravaDocResultado}
\alias{gravaDocResultado}
\title{Records doc result}
\usage{
gravaDocResultado(dfResultado, arquivo, template)
}
\arguments{
\item{dfResultado}{a dataset to record in doc}

\item{arquivo}{the name that you want to file}

\item{template}{"character" value, it represents the filename of the docx file used as a template.}
}
\description{
this function records the result in a docx file
}

