% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravaResultado.R
\name{gravaResultado}
\alias{gravaResultado}
\title{Records doc result}
\usage{
gravaResultado(resultado, arquivo, modelo, template = "template.docx")
}
\arguments{
\item{resultado}{table with the results of statistical functions available in avaliaAjuste function}

\item{arquivo}{the name that you want to file}

\item{modelo}{Rdata file}

\item{template}{"character" value, it represents the filename of the docx file used as a template.}
}
\description{
this function records the result in a docx file
}

