% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ifrm.R
\name{ifrm}
\alias{ifrm}
\title{ifrm}
\usage{
ifrm(obj, env = globalenv())
}
\arguments{
\item{obj}{the object that you want to remove}

\item{env}{The global environment}
}
\description{
if the object does not exist an error will not happen.
}
\examples{
a = 5
ifrm(a)
ifrm(b)
}

