% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isfinitedataframe.R
\name{isfinitedataframe}
\alias{isfinitedataframe}
\title{is finite data frame}
\usage{
isfinitedataframe(obj)
}
\arguments{
\item{obj}{any object}
}
\value{
TRUE if "x" is finite, FALSE if "x" is not finite
}
\description{
check if a data.frame has any non-finite elements
}
\examples{
date <- c("02/2009","02/2010","02/2011","02/2012")
x <- c(1,2,3,4)
test <- data.frame(x,date)
isfinitedataframe(test)
isfinitedataframe(x)
}

