% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listToDataFrame.R
\name{listToDataFrame}
\alias{listToDataFrame}
\title{List to DataFrame}
\usage{
listToDataFrame(dlist)
}
\arguments{
\item{dlist}{a list}
}
\description{
converts a list in a dataframe
}
\examples{
a <- 1:5
listToDataFrame(a)
b = listToDataFrame(a)
}

