% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mae.R
\name{mae}
\alias{mae}
\title{mean absolute error (mae)}
\usage{
mae(observados, estimados)
}
\arguments{
\item{observados}{vector of values observed.}

\item{estimados}{vector of regression model data.}
}
\value{
Function that returns Mean Absolute Error
}
\description{
is a quantity used to measure how close forecasts or predictions are to the eventual outcomes. The mean absolute error is given by.
}
\details{
mae = mean(abs(observados-estimados))
}
\references{
see \url{https://en.wikipedia.org/wiki/Mean_absolute_error} for more details.
}

