% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mspr.R
\name{mspr}
\alias{mspr}
\title{mspr}
\usage{
mspr(observados, estimados, nValidacao)
}
\arguments{
\item{observados}{vector of values observed.}

\item{estimados}{vector of regression model data.}

\item{nValidacao}{number of cases in the validation data set.}
}
\description{
average square of the prediction errors .
}
\references{
JESUS, S. C.; MIURA, A. K. Analise de regressao linear multipla para estimativa do indice de vegetacao melhorado (EVI) a partir das bandas 3 4 e 5 do sensor TM/Landsat 5. In: SIMPOSIO BRASILEIRO DE SENSORIAMENTO REMOTO, 14. (SBSR), 2009, Natal. Anais... Sao Jose dos Campos: INPE, 2009. p. 1103-1110. DVD, On-line. ISBN 978-85-17-00044-7. (INPE-15901-PRE/10511)
}

