% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predizer.R
\name{predizer}
\alias{predizer}
\title{Predict}
\usage{
predizer(ajuste, newdata, force = FALSE, ...)
}
\arguments{
\item{ajuste}{is ajust obtained a function like lm or nlsLM}

\item{newdata}{dataframe where fields will be update}

\item{force}{force the calculation without using predict?}

\item{...}{only for compatibility with other functions}
}
\value{
will be returned list of values predicts
}
\description{
this function is the replacement predict, she tries to predict if the return zero predict it calculates the prediction with the coefficients reported in the parameter setting
}
