% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectBaseOriented.R
\name{projectBaseOriented}
\alias{projectBaseOriented}
\title{Project Base Oriented}
\usage{
projectBaseOriented(firstAge = NaN, lastAge = NaN, fitDAP, fitHT, base,
  mapper = list(age1 = "idadearred1", dap1 = "dap1", dap2 = "dap2", ht1 =
  "ht1", ht2 = "ht2"), calcVolume = calculaVolumeDefault, forcePredict = F)
}
\arguments{
\item{firstAge}{the first age to predict}

\item{lastAge}{the last age to predict}

\item{fitDAP}{a fit get function inherit lm to DAP}

\item{fitHT}{a fit get function inherit lm to HT}

\item{base}{data base}

\item{mapper}{the label used in fields to age, dap and ht}

\item{calcVolume}{function to calc volume}

\item{forcePredict}{force calc base coefficients or se predict()?}
}
\value{
will be returned a list of volume predict to ages in dataframe and/or param
}
\description{
this function build a list of dataframe with projects of ages between 'firstAge' and 'lastAge' params
}

