% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R21a.R
\name{R21a}
\alias{R21a}
\title{R21a}
\usage{
R21a(observados, estimados, k)
}
\arguments{
\item{observados}{vector of values observed.}

\item{estimados}{vector of values estimated.}

\item{k}{is the number of model parameters}
}
\description{
To avoid any problems and confudion on the part of the data analyst, it seems a safe recommendation to use R21a consistently for any type of model with the appropeiate a value, rather than ajusting any of the other
}
\details{
R21a <- 1-a*(1 - R21)
}

