% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R29a.R
\name{R29a}
\alias{R29a}
\title{R29a}
\usage{
R29a(observados, estimados, k)
}
\arguments{
\item{observados}{vector of values observed.}

\item{estimados}{vector of values estimated.}

\item{k}{is the number of model parameters}
}
\description{
To avoid any problems and confusion on the part of the data analyst, it seems a safe recommendation to use R21a consistently for any type of model with the appropeiate a value, rather than adjusting any of the other.
}
\details{
R29a <- 1-a*(1 - R29)
}
