% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/residuoPerc.R
\name{residuoPerc}
\alias{residuoPerc}
\title{calculates residue percentage}
\usage{
residuoPerc(observados, estimados)
}
\arguments{
\item{observados}{vector of values observed.}

\item{estimados}{vector of values estimated.}
}
\description{
this function calculates the vector residue percentage.
}
\details{
calculaPerc = ((valor)/mean(observados))*100
}

