% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retornaValor.R
\name{retornaValor}
\alias{retornaValor}
\title{return value}
\usage{
retornaValor(valor)
}
\arguments{
\item{valor}{any variable}
}
\value{
the variable converted to its value
}
\description{
this feature is designed to fix variables that its content was a command
}
\examples{
a = 5
retornaValor(a)
}
