% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{Root Mean Square Error}
\usage{
rmse(observados, estimados)
}
\arguments{
\item{observados}{vector of values observed.}

\item{estimados}{vector of regression model data.}
}
\description{
The root-mean-square error (RMSE) is a frequently used measure of the differences between values (sample and population values) predicted by a model or an estimator and the values actually observed.
}
\details{
rmse = sqrt(mean((observados - estimados)^2))
}
\references{
See \url{https://en.wikipedia.org/wiki/Root-mean-square_deviation} for more details.
}

