% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundAge.R
\name{roundAge}
\alias{roundAge}
\title{Round Ages}
\usage{
roundAge(plots, ages, inYears = F, firstAge = NaN)
}
\arguments{
\item{plots}{is list of plots}

\item{ages}{is list of age}

\item{inYears}{ages are in year?}

\item{firstAge}{synchronize begin of ages with an age? what age?}
}
\value{
will be returned a list of round ages
}
\description{
this function approaching the age to the nearest age as an integer
}
