% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrmse.R
\name{rrmse}
\alias{rrmse}
\title{relative root mean square error}
\usage{
rrmse(observados, estimados)
}
\arguments{
\item{observados}{vector of values observed.}

\item{estimados}{vector of regression model data.}
}
\description{
relative root mean square error (RRMSE) is calculated by dividing the RMSE by the mean observed data
}
