% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salvaModelo.R
\name{salvaModelo}
\alias{salvaModelo}
\title{save function with Model}
\usage{
salvaModelo(modelo, diretorio = "")
}
\arguments{
\item{modelo}{function with Model the save}

\item{diretorio}{directory to save the file, if not informed saved in the work directory}
}
\description{
save function with Model of type criaModeloGenerico or criaModeloExclusivo
}

