% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separaDados.R
\name{separaDados}
\alias{separaDados}
\title{Data Separates}
\usage{
separaDados(dataFrame, fieldName, percTraining = 0.7, seed = NULL)
}
\arguments{
\item{dataFrame}{source of data}

\item{fieldName}{column of dataFrame that will be applied regression}

\item{percTraining}{percentage that will be reserved for training (default 0.70)}

\item{seed}{integer that determines how the sample is randomly chosen (default NULL)}
}
\description{
divides the dataFrame as the percentage defined in percTraining enabling apply and measure the performance of the regression equation.
}

