% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syx.R
\name{syx}
\alias{syx}
\title{Standard Error of Estimate}
\usage{
syx(observados, estimados, n, p)
}
\arguments{
\item{observados}{vector of values observed.}

\item{estimados}{vector of values estimated.}

\item{n}{the amount of values observed}

\item{p}{the size of the vector of regression model data}
}
\description{
Measures the variability, or scatter of the observed values around the regression line
}
