% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syxPerc.R
\name{syxPerc}
\alias{syxPerc}
\title{Standard Error of Estimate Percentage}
\usage{
syxPerc(syx, observados)
}
\arguments{
\item{syx}{result of the function syx(Standard Error of Estimate).}

\item{observados}{vector of values observed.}
}
\description{
Measures the variability, or scatter of the observed values around the regression line
}

