% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichmedian.R
\name{whichmedian}
\alias{whichmedian}
\title{whichmedian}
\usage{
whichmedian(x)
}
\arguments{
\item{x}{a vector of numbers}
}
\value{
vector position that has its closest median value
}
\description{
vector position that has its closest median value
}
\examples{
dados <- c(1,2,3,4,9,5,6)
whichmedian(dados)
}

