% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificaClasseDAP.R
\name{classificaClasseDAP}
\alias{classificaClasseDAP}
\title{classifica Classe DAP}
\usage{
classificaClasseDAP(dfClassesDAP, dap)
}
\arguments{
\item{dfClassesDAP}{a frequency distribution with the attributes $classe and $centro}

\item{dap}{integer Diameter at breast height}
}
\description{
the center of the class that the DAP belongs.
}
\examples{
dados <- c(1,2,3,4)
dados = defineClasses2(dados,2)
classificaClasseDAP(dados,2)
}

