% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/converteCampoParaCharacter.R
\name{converteCampoParaCharacter}
\alias{converteCampoParaCharacter}
\title{Field Converts To Character}
\usage{
converteCampoParaCharacter(nomeCampo, base)
}
\arguments{
\item{nomeCampo}{the column name you want to convert}

\item{base}{the column having dataFrame, that you want to convert to String}
}
\value{
base dataFrame with a column converted to String
}
\description{
converts a column of a dataframe to String
}
\examples{
measurement_date <- c(02/2009,02/2010,02/2011,02/2011)
plot <- c(1,2,3,4)
test <- data.frame(measurement_date,plot)
converteCampoParaCharacter("measurement_date",test)
}

