% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineClasses.R
\name{defineClasses}
\alias{defineClasses}
\title{define Classes}
\usage{
defineClasses(limiteMin, limiteMax, amplitude, decrescente = TRUE)
}
\arguments{
\item{limiteMin}{the lowest list number}

\item{limiteMax}{the largest number in the list}

\item{amplitude}{List amplitude}

\item{decrescente}{order by true decreasing , false increasing}
}
\description{
creates a list with the class interval of a frequency distribution
}

