% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getAnoMedicao.R
\name{getAnoMedicao}
\alias{getAnoMedicao}
\title{Get Year Measurement}
\usage{
getAnoMedicao(dataFrame, column_name_measurement_date, column_name_plot)
}
\arguments{
\item{dataFrame}{that has the column DATE(MM/YYYY) and a ID column_name_plot}

\item{column_name_measurement_date}{column with a date format}

\item{column_name_plot}{a column of dataFrame, identification of plot (ID_plot)}
}
\value{
dataFrame dataframe that has columns column_name_measurement_date, column_name_plot, ANO_MEDICAO
}
\description{
using column_name_measurement_date column in the form MM/YYYY creates a new column with the name "ANO_MEDICAO" in YYYY format
}
\examples{
column_name_measurement_date <- c("02/2009","02/2010","02/2011","02/2012")
column_name_plot <- c(1,2,3,4)
test <- data.frame(column_name_measurement_date,column_name_plot)
getAnoMedicao(test,"column_name_measurement_date","column_name_plot")
}

