% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/verificaTipoColuna.R
\name{verificaTipoColuna}
\alias{verificaTipoColuna}
\title{Check de type of Column}
\usage{
verificaTipoColuna(coluna)
}
\arguments{
\item{coluna}{column of dataframe}
}
\description{
this function returns the type of a column of a dataFrame, if it is numeric or character.
}
\examples{
ID_REGIAO <- c(1,2,3,4)
CD_PLANTIO <- c("ACD","CDB","CDC","CDD")
test <- data.frame(ID_REGIAO,CD_PLANTIO)
verificaTipoColuna(test$ID_REGIAO)
}

